requirements.txt
easyocr==1.7.2
pillow==11.1.0
Pyrogram==2.0.106
requests==2.32.3
beautifulsoup4==4.12.3